requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
aiohttp==3.12.15
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
jinja2==3.1.6
python-dotenv==1.1.1
aiosmtplib==4.0.2
email-validator==2.3.0
pytest==8.4.1
pytest-asyncio==1.1.0